Breaking out job queue to its own file
