+++ src/server/server.ts
@@ -1,14 +1,12 @@
 import * as bodyParser from "body-parser"
-import * as spawn from "child_process"
 import * as dotenv from "dotenv"
 import * as express from "express"
 import * as fs from "fs"
-import * as Pq from "p-queue"
 import * as path from "path"
 import * as Raven from "raven"
 import * as winston from "winston"
-
 import * as Hub from "../hub"
+import * as AsyncProcessQueue from "./action_process_queue"
 import * as apiKey from "./api_key"
 
 const expressWinston = require("express-winston")
@@ -18,7 +16,7 @@ const TOKEN_REGEX = new RegExp(/[T|t]oken token="(.*)"/)
 const statusJsonPath = path.resolve(`${__dirname}/../../status.json`)
 const useRaven = () => !!process.env.ACTION_HUB_RAVEN_DSN
 
-const queue = new Pq({concurrency: 1})
+const expensiveJobQueue = new AsyncProcessQueue.AsyncProcessJob()
 
 export default class Server implements Hub.RouteBuilder {
 
@@ -117,7 +115,7 @@ export default class Server implements Hub.RouteBuilder {
       const request = Hub.ActionRequest.fromRequest(req)
       const action = await Hub.findAction(req.params.actionId, { lookerVersion: request.lookerVersion })
       if (action.expensive && action.usesStreaming) {
-        await queue.add(async () => this.asyncProcess(req, res))
+        await expensiveJobQueue.asyncProcess(req, res)
       } else {
         const actionResponse = await action.validateAndExecute(request)
 
@@ -223,43 +221,6 @@ export default class Server implements Hub.RouteBuilder {
       instanceId: req.header("x-looker-instance"),
       webhookId: req.header("x-looker-webhook-id"),
     }
-  }
-
-  private async asyncProcess(req: express.Request, res: express.Response) {
-      return new Promise<void>((resolve, reject) => {
-          const child = spawn.fork(`./src/actions/actions_process.ts`)
-          child.on("message", (actionResponse) => {
-              // Some versions of Looker do not look at the "success" value in the response
-              // if the action returns a 200 status code, even though the Action API specs otherwise.
-              // So we force a non-200 status code as a workaround.
-              winston.info("Success? " + actionResponse.success + " || and what: " + JSON.stringify(actionResponse))
-              if (!actionResponse.success) {
-                  res.status(400)
-              }
-              res.json(actionResponse)
-
-              child.kill()
-              resolve()
-          }).on("error", (err) => {
-            winston.warn(err.message)
-            if (!child.killed) {
-              child.kill()
-            }
-            reject(err)
-          }).on("exit", () => {
-              if (!child.killed) {
-                  child.kill()
-              }
-          })
-          const data = {
-              body: req.body,
-              actionId: req.params.actionId,
-              instanceId: req.header("x-looker-instance"),
-              webhookId: req.header("x-looker-webhook-id"),
-              userAgent: req.header("user-agent"),
-          }
-          child.send(data)
-      })
   }
 
   private absUrl(rootRelativeUrl: string) {
